add kafka announcement only logic
