+++ gradle.properties
@@ -1,4 +1,4 @@
-version=29.58.11
+version=29.58.12
 group=com.linkedin.pegasus
 org.gradle.configureondemand=true
 org.gradle.parallel=true
